app/utils/requirements.txt
st-pages==0.4.5
streamlit==1.30
streamlit-feedback==0.1.2
requests
pyyaml
pypdf2
openai>=1.32.0,<2.0.0

tqdm
tiktoken
psycopg2-binary
sqlalchemy

pymysql
altair==4.0
geopy

unidecode
cryptography
langchain
langchain-openai

fastapi
requests
websockets